chore: refactor setup credentials
Refactors the `setupCredentials` method from `test/sqlamdin-fetcher.ts`
into a reusable test helper module.
